+++ example/lib/main.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_widget_function/function/http_ssl_certificate.dart';
 import 'package:flutter_widget_function/widget/keyboard/keyboard_dismiss.dart';
 import 'package:flutter_widget_function_example/dashboard.dart';
 import 'package:flutter_widget_function_example/routes/routes.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
+  HttpSSLCertificate.disable();
   runApp(const MyApp());
 }